Color Keqing

+++ include/Keqing.hpp
@@ -62,7 +62,7 @@ public:
 
     static inline Keqing *getInstance() { return instance; }
 
-    void colorTexture(int r, int g, int b);
+    void colorTexture(Uint8 r, Uint8 g, Uint8 b);
 
     void moveX() override;
